// CourseEnrollmentApplication/src/app/model/courses.ts
export class Courses {
  course_id: string | undefined;
  course_name: string | undefined;
  course_author: string | undefined;
  course_category: string | undefined;
  course_creationDate: string | undefined;
  course_authorExp: string | undefined;
  status: string | undefined;

  constructor() {}
}